requirements.txt
Django==4.2.7
django-allauth==0.59.0
django-ckeditor==6.4.2
django-cloudinary-storage==0.3.0
django-hitcount==1.3.5
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-mptt==0.16.0
pillow
python-decouple==3.8
django-formtools
